requirements.txt
flask==2.1.3
flask-restx==1.2.0
Flask-Cors==4.0.0
Werkzeug==2.0.3
loguru==0.4.1
python-dotenv==1.0.0
numpy==1.24.3
protobuf==4.21.12
gunicorn==19.10.0
requests
pytest